Tue, Apr 26, 2022, 3:34 PM +00:00

+++ week3/runoff/runoff.c
@@ -166,7 +166,7 @@ bool print_winner(void)
 {
     for (int i = 0; i < candidate_count; i++)
     {
-        if ((candidate[i].votes / voter_count) > 0.50)
+        if ((float)(candidate[i].votes / voter_count) > 0.50)
         {
             printf("%s\n", candidate[i].name);
             return true;
